refactor!: move SurveyConstants to its own package `domain.surveys`
