+++ musical-surveyor-app/src/main/java/io/davorpatech/apps/musicalsurveyor/domain/surveys/SuveyStatus.java
@@ -1,4 +1,4 @@
-package io.davorpatech.apps.musicalsurveyor.domain;
+package io.davorpatech.apps.musicalsurveyor.domain.surveys;
 
 /**
  * Enumeration of the possible states of a survey.
