+++ musical-surveyor-app/src/main/java/io/davorpatech/apps/musicalsurveyor/persistence/model/Survey.java
@@ -1,7 +1,7 @@
 package io.davorpatech.apps.musicalsurveyor.persistence.model;
 
-import io.davorpatech.apps.musicalsurveyor.domain.SurveyConstants;
-import io.davorpatech.apps.musicalsurveyor.domain.SuveyStatus;
+import io.davorpatech.apps.musicalsurveyor.domain.surveys.SurveyConstants;
+import io.davorpatech.apps.musicalsurveyor.domain.surveys.SuveyStatus;
 import io.davorpatech.fwk.auditing.jpa.Audit;
 import io.davorpatech.fwk.auditing.jpa.AuditAccessor;
 import io.davorpatech.fwk.model.BaseEntity;
