+++ musical-surveyor-app/src/main/java/io/davorpatech/apps/musicalsurveyor/persistence/model/SurveyParticipation.java
@@ -1,6 +1,6 @@
 package io.davorpatech.apps.musicalsurveyor.persistence.model;
 
-import io.davorpatech.apps.musicalsurveyor.domain.SurveyParticipationConstants;
+import io.davorpatech.apps.musicalsurveyor.domain.surveys.SurveyParticipationConstants;
 import io.davorpatech.fwk.auditing.jpa.Audit;
 import io.davorpatech.fwk.auditing.jpa.AuditAccessor;
 import io.davorpatech.fwk.model.BaseEntity;
